backend/requirements.txt
absl-py==2.1.0
accelerate==1.8.1
agentic-doc==0.2.7
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.4.1
alyahmor==0.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.3
anyio==4.9.0
arabic-reshaper==3.0.0
Arabic-Stopwords==0.4.3
ArabicOcr==1.1.6
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arramooz-pysqlite==0.4.2
arrow==1.3.0
asgiref==3.8.1
asmai==0.1
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
av==14.4.0
azure-ai-documentintelligence==1.0.2
azure-ai-inference==1.0.0b9
azure-core==1.35.1
azure-storage-blob==12.26.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bitsandbytes==0.46.0
bleach==6.2.0
blinker==1.8.2
boto3==1.38.37
botocore==1.38.37
build==1.2.2.post1
cachetools==5.5.2
camel_tools==1.5.6
catalogue==2.0.10
cattrs==25.1.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.12
click==8.2.1
codernitydb3==0.6.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
coremltools==8.3.0
courlan==1.3.2
cryptography==45.0.4
cycler==0.12.1
dataclasses-json==0.6.7
dateparser==1.2.2
debugpy==1.8.14
decorator==5.2.1
deepdoctection==0.43.2
defusedxml==0.7.1
Deprecated==1.2.18
diffusers==0.10.0
dill==0.4.0
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
docx2pdf==0.1.8
durationpy==0.10
easyocr==1.7.2
editdistance==0.8.1
editdistpy==0.1.6
effdet==0.4.1
email_validator==2.2.0
emoji==2.14.1
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers==24.3.25
fonttools==4.55.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.5.1
ftfy==6.3.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud==0.34.0
google-cloud-vision==3.10.2
google-generativeai==0.8.5
google_search_results==2.4.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
h5py==3.14.0
htmldate==1.9.3
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
img2pdf==0.6.1
importlib_metadata==8.4.0
importlib_resources==6.5.2
iopath==0.1.10
ipykernel==6.30.1
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.2
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.31
jaxlib==0.4.31
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
json5==0.12.0
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
jusText==3.0.2
kiwisolver==1.4.7
kraken==5.3.0
kubernetes==33.1.0
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.67
langchain-openai==0.3.23
langchain-text-splitters==0.3.8
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.3.45
language_data==1.3.0
layoutparser==0.3.4
lazy-imports==0.3.1
lazy_loader==0.4
Levenshtein==0.27.1
libqutrub==1.2.4.1
libtpu-nightly==0.1.dev20240729
lightning==2.4.0
lightning-utilities==0.14.3
locate==1.1.1
lxml==5.4.0
lxml_html_clean==0.4.2
marisa-trie==1.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
maskouk-pysqlite==0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.18
mishkal==0.4.1
mistune==3.1.3
ml-dtypes==0.4.0
mmh3==5.1.0
mock==4.0.3
mpmath==1.3.0
mrjob==0.7.4
msgpack==1.1.0
muddler==0.1.3
multidict==6.4.4
mypy_extensions==1.1.0
mysam-tagmanager==0.4.1
Naftawayh==0.4
narwhals==1.9.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
ninja==1.11.1.4
notebook_shim==0.2.4
npm==0.1.1
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.0
openai==1.109.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
opt-einsum==3.3.0
optional-django==0.1.0
orjson==3.11.0
overrides==7.7.0
packaging==24.2
paddlepaddle==3.0.0
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pickleDB==1.3.2
pikepdf==9.9.0
pillow==11.2.1
pillow_heif==0.22.0
platformdirs==4.3.8
portalocker==3.2.0
posthog==5.0.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyaml==25.5.0
PyArabic==0.6.15
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.26.1
pyparsing==3.2.3
pypdf==5.6.0
PyPDF2==3.0.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyrsistent==0.20.0
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-doctr==0.12.0
python-docx==1.2.0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-Levenshtein==0.27.1
python-multipart==0.0.20
pytorch-lightning==2.5.1.post0
pytz==2025.2
pyvips==3.0.0
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.0
qalsadi==0.5.1
qwen-vl-utils==0.0.11
RapidFuzz==3.13.0
referencing==0.35.1
regex==2024.11.6
reportlab==4.4.2
requests==2.32.4
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.2
rpds-py==0.20.0
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.4.2
scipy==1.13.1
Send2Trash==1.8.3
sentence-transformers==2.6.1
sentencepiece==0.2.0
sentry-sdk==2.33.0
setuptools==80.9.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.47.1
streamlit==1.39.0
structlog==25.4.0
sylajone==0.3.1
sympy==1.13.1
symspellpy==6.9.0
tabulate==0.9.0
Tashaphyne==0.3.6
tenacity==9.0.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2025.6.11
tiktoken==0.9.0
timm==1.0.15
tinycss2==1.4.0
tld==0.13.1
tldextract==5.3.0
tokenizers==0.21.2
toml==0.10.2
torch==2.6.0.dev20241112+cu121
torchaudio==2.5.0.dev20241112+cu121
torchmetrics==1.7.3
torchvision==0.20.0.dev20241112+cu121
tornado==6.4.1
tqdm==4.67.1
trafilatura==2.0.0
traitlets==5.14.3
transformers==4.52.4
typer==0.16.0
types-python-dateutil==2.9.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
validators==0.35.0
watchdog==5.0.3
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
widgetsnbextension==4.0.14
wordfreq==3.1.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.20.1
zstandard==0.23.0
